requirements.txt
streamlit
python-dotenv
pymongo